Upload of list property webpage
This commit includes the design of the list property webpage and its logic to make sure things such as the availability calendar function as required. Changes will need to be made to improve upon this design and to make sure that the availability set on the calendar is submitted by the webpage as well.