Improvements made to list property page
List property page now has multiple inputs to for every separate address line making the inputs easier to understand, preventing misunderstandings. The price, number of bedrooms and number of bathrooms inputs have been changed to each have a minimum value to prevent issues caused by this. Along with this the price has been changed to have a step of 0.01 which makes sense for a price and prevents issues with incorrect values being entered. The arrows for the number inputs have been removed as well for an improved design. Along with this a JavaScript function has been added to prevent the number inputs from changing when active and scrolled upon.